% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_choc.R
\docType{data}
\name{data_choc}
\alias{data_choc}
\title{This is documentation}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 180 rows and 9 columns.
}
\usage{
data_choc
}
\description{

}
\keyword{datasets}
